// src/communaute/nouvelle_ame/interfaces.ts
export interface INouvelleAme {
	idNouvelleAme: number;
	nomNouvelleAme: string;
	prenomNouvelleAme: string;
	dateConversionNouvelleAme: string;
	lieuHabitationNouvelleAme: string;
	fonctionNouvelleAme: string;
	contactNouvelleAme: string;
	sexeNouvelleAme: string;
	emailNouvelleAme: string;
	idEglise: number;
	idCellule: number;
	idDepartement: number;
	idResponsabilite: number;
	idDomaineActivite: number;
	idUtilisateur: number;
	idGroupe: number;
}